00_metadata_raglite_summarization/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "metadata-rag-lite"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "pydantic>=2.6",
  "jsonschema>=4.21",
  "python-dotenv>=1.0",
]

[project.scripts]
metaragl = "metaragl.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]